+++ src/lib/satteri-external-links.mjs
@@ -1,13 +1,10 @@
 /**
  * Sätteri HAST plugin to open external links in a new tab.
- * (Replaces `rehype-external-links`.)
  *
  * Adds `target="_blank"` + `rel="noopener noreferrer"` to links that leave
  * the site. "External" deliberately mirrors SmartLink's semantics
  * (`src/components/mdx/SmartLink.astro`): the href starts with `http` and
- * does not contain `danny.is`. This is a small improvement over the old
- * `rehype-external-links` config, which blank-targeted *all* absolute URLs
- * including danny.is ones.
+ * does not contain `danny.is`.
  *
  * In practice this only affects plain `.md` content — in `.mdx` the
  * `a -> SmartLink` remapping computes the same attributes itself (the
